refactor: renamed file and class name
