+++ app/services/contribution_services/fees_labeling.rb
@@ -1,29 +1,32 @@
-module Contributions
-  class RemainingContributionFeeHandlerService
-    attr_reader :contribution, :initial_contributions_balance, :remaining_fee
+module ContributionServices
+  class FeesLabeling
+    attr_reader :contribution, :initial_contributions_balance
 
     CONTRACT_FEE_PERCENTAGE = 0.1
 
-    def initialize(contribution:, remaining_fee:)
+    def initialize(contribution:)
       @contribution = contribution
-      @remaining_fee = remaining_fee
-      @initial_contributions_balance = ContributionBalance.sum(:tickets_balance_cents)
+      @initial_contributions_balance = ContributionBalance.sum(:fees_balance_cents)
     end
 
-    def spread_remaining_fee
-      return if not_enough_tickets_balance?
-
+    def spread_fee_to_payers
+      deal_with_fees_balances_empty
       create_fees_for_feeable_contributions
+    rescue StandardError => e
+      Reporter.log(error: e)
     end
 
     private
 
-    def not_enough_tickets_balance?
-      feeable_contribution_balances.empty?
+    def deal_with_fees_balances_empty
+      return unless feeable_contribution_balances.empty?
+
+      RemainingContributionFeeHandler
+        .new(contribution:, remaining_fee: fee_generated_by_new_contribution).spread_remaining_fee
     end
 
     def create_fees_for_feeable_contributions
-      accumulated_fees_result = remaining_fee.ceil
+      accumulated_fees_result = fee_generated_by_new_contribution.ceil
 
       feeable_contribution_balances.each do |contribution_balance|
         if last_payer?(accumulated_fees_result:, contribution_balance:)
@@ -39,16 +42,8 @@ module Contributions
     end
 
     def handle_fee_creation_for(contribution_balance:, fee_cents:, contribution_increased_amount_cents:)
-      transfer_ticket_balance_to_fees_balance(contribution_balance:, fee_cents:)
-
       ContributionFeeCreatorService.new(contribution_balance:, fee_cents:, contribution:,
                                         contribution_increased_amount_cents:).handle_fee_creation
-    end
-
-    def transfer_ticket_balance_to_fees_balance(contribution_balance:, fee_cents:)
-      contribution_balance.tickets_balance_cents -= fee_cents
-      contribution_balance.fees_balance_cents += fee_cents
-      contribution_balance.save
     end
 
     def last_payer?(accumulated_fees_result:, contribution_balance:)
@@ -59,28 +54,28 @@ module Contributions
       RibonConfig.minimum_contribution_chargeable_fee_cents
     end
 
+    def fee_generated_by_new_contribution
+      contribution.generated_fee_cents
+    end
+
     def feeable_contribution_balances
       @feeable_contribution_balances ||= ContributionQueries.new(contribution:)
-                                                            .ordered_feeable_tickets_contribution_balances
+                                                            .ordered_feeable_contribution_balances
     end
 
     def fee_and_increased_value_for(contribution_balance:)
-      payer_balance = contribution_balance.tickets_balance_cents
-      fee_to_be_paid = remaining_fee
+      payer_balance = contribution_balance.fees_balance_cents
+      fee_to_be_paid = fee_generated_by_new_contribution
 
-      ContributionFeeCalculatorService
+      ContributionFeeCalculator
         .new(payer_balance:, fee_to_be_paid:, initial_contributions_balance:)
         .fee_and_increased_value_for(contribution:)
     end
 
     def handle_last_contribution_fee(accumulated_fees_result:, contribution_balance:)
-      # TODO: refactor this logic to use in last_contribution_fee_handler_service
-
-      fee_cents = [accumulated_fees_result, contribution_balance.tickets_balance_cents].min
-      contribution_increased_amount_cents =
-        contribution.usd_value_cents * fee_cents / contribution.generated_fee_cents.to_f
-
-      handle_fee_creation_for(contribution_balance:, fee_cents:, contribution_increased_amount_cents:)
+      LastContributionFeeHandler
+        .new(accumulated_fees_result:, contribution_balance:, contribution:)
+        .charge_remaining_fee
     end
   end
 end
