+++ app/services/contribution_services/remaining_contribution_fee_handler.rb
@@ -1,32 +1,29 @@
-module Contributions
-  class FeesLabelingService
-    attr_reader :contribution, :initial_contributions_balance
+module ContributionServices
+  class RemainingContributionFeeHandler
+    attr_reader :contribution, :initial_contributions_balance, :remaining_fee
 
     CONTRACT_FEE_PERCENTAGE = 0.1
 
-    def initialize(contribution:)
+    def initialize(contribution:, remaining_fee:)
       @contribution = contribution
-      @initial_contributions_balance = ContributionBalance.sum(:fees_balance_cents)
+      @remaining_fee = remaining_fee
+      @initial_contributions_balance = ContributionBalance.sum(:tickets_balance_cents)
     end
 
-    def spread_fee_to_payers
-      deal_with_fees_balances_empty
+    def spread_remaining_fee
+      return if not_enough_tickets_balance?
+
       create_fees_for_feeable_contributions
-    rescue StandardError => e
-      Reporter.log(error: e)
     end
 
     private
 
-    def deal_with_fees_balances_empty
-      return unless feeable_contribution_balances.empty?
-
-      RemainingContributionFeeHandlerService
-        .new(contribution:, remaining_fee: fee_generated_by_new_contribution).spread_remaining_fee
+    def not_enough_tickets_balance?
+      feeable_contribution_balances.empty?
     end
 
     def create_fees_for_feeable_contributions
-      accumulated_fees_result = fee_generated_by_new_contribution.ceil
+      accumulated_fees_result = remaining_fee.ceil
 
       feeable_contribution_balances.each do |contribution_balance|
         if last_payer?(accumulated_fees_result:, contribution_balance:)
@@ -42,8 +39,16 @@ module Contributions
     end
 
     def handle_fee_creation_for(contribution_balance:, fee_cents:, contribution_increased_amount_cents:)
+      transfer_ticket_balance_to_fees_balance(contribution_balance:, fee_cents:)
+
       ContributionFeeCreatorService.new(contribution_balance:, fee_cents:, contribution:,
                                         contribution_increased_amount_cents:).handle_fee_creation
+    end
+
+    def transfer_ticket_balance_to_fees_balance(contribution_balance:, fee_cents:)
+      contribution_balance.tickets_balance_cents -= fee_cents
+      contribution_balance.fees_balance_cents += fee_cents
+      contribution_balance.save
     end
 
     def last_payer?(accumulated_fees_result:, contribution_balance:)
@@ -54,28 +59,28 @@ module Contributions
       RibonConfig.minimum_contribution_chargeable_fee_cents
     end
 
-    def fee_generated_by_new_contribution
-      contribution.generated_fee_cents
-    end
-
     def feeable_contribution_balances
       @feeable_contribution_balances ||= ContributionQueries.new(contribution:)
-                                                            .ordered_feeable_contribution_balances
+                                                            .ordered_feeable_tickets_contribution_balances
     end
 
     def fee_and_increased_value_for(contribution_balance:)
-      payer_balance = contribution_balance.fees_balance_cents
-      fee_to_be_paid = fee_generated_by_new_contribution
+      payer_balance = contribution_balance.tickets_balance_cents
+      fee_to_be_paid = remaining_fee
 
-      ContributionFeeCalculatorService
+      ContributionFeeCalculator
         .new(payer_balance:, fee_to_be_paid:, initial_contributions_balance:)
         .fee_and_increased_value_for(contribution:)
     end
 
     def handle_last_contribution_fee(accumulated_fees_result:, contribution_balance:)
-      LastContributionFeeHandlerService
-        .new(accumulated_fees_result:, contribution_balance:, contribution:)
-        .charge_remaining_fee
+      # TODO: refactor this logic to use in last_contribution_fee_handler_
+
+      fee_cents = [accumulated_fees_result, contribution_balance.tickets_balance_cents].min
+      contribution_increased_amount_cents =
+        contribution.usd_value_cents * fee_cents / contribution.generated_fee_cents.to_f
+
+      handle_fee_creation_for(contribution_balance:, fee_cents:, contribution_increased_amount_cents:)
     end
   end
 end
